pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "wheel"]

[project]
name = "OptiMates"
description = "Tracking pipeline using motile to find trajectories and training linajea to predict detections and movement vectors."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
  "Programming Language :: Python :: 3",
]
keywords = []
license = { text = "BSD 3-Clause License" }
authors = [
    { email = "chiaralo@ethz.ch", name = "Chiara Lombardo" },
]
dynamic = ["version"]
dependencies = [
  "LineageTree @git+https://github.com/BadPrograms/LineageTree-1.git",
  "funlib.persistence @git+https://github.com/funkelab/funlib.persistence.git",
  "funlib.learn.torch @git+https://github.com/funkelab/funlib.learn.torch.git",
  "gunpowder",
  "daisy",
  "torch",
  "linajea @git+https://github.com/funkelab/linajea.git",
]

[project.optional-dependencies]
dev = [
  'pytest',
  'pytest-cov',
  'black',
  'mypy',
  'pdoc',
  'pre-commit'
]

[project.urls]
homepage = "https://github.com/chiaralooo/OptiMates"
repository = "https://github.com/chiaralooo/OptiMates"
